Move component into folder
