+++ ui/src/shared/components/group_by_time_dropdown/GroupByTimeDropdown.tsx
